fix(fr-command/store): redis.REPL_AOF / REPL_SLAVE constants match upstream encoding [bd-frankenredis-replconst]
Upstream server.h:
  #define PROPAGATE_AOF  1
  #define PROPAGATE_REPL 2
script_lua.c then exports redis.REPL_AOF = PROPAGATE_AOF and
REPL_SLAVE = REPL_REPLICA = PROPAGATE_REPL.

fr had the values swapped:
  REPL_AOF = 2.0
  REPL_SLAVE / REPL_REPLICA = 1.0
And fr-store's SCRIPT_PROPAGATE_AOF / SCRIPT_PROPAGATE_REPLICA
used the same swap, so redis.set_repl(redis.REPL_AOF) toggled the
replica bit instead of the AOF bit when scripts changed
propagation per-call.

Fix:
- fr-store: swap SCRIPT_PROPAGATE_AOF (now 0b0001) and
  SCRIPT_PROPAGATE_REPLICA (now 0b0010).
- fr-command/lua: swap REPL_AOF (now 1.0) and REPL_SLAVE /
  REPL_REPLICA (now 2.0).
- Update eval_tracks_script_propagation_targets_across_set_repl_changes
  to expect REPL_REPLICA + REPL_SLAVE = 4.

Differential probe vs vendored redis 7.2.4 byte-matches all four
constant readbacks.

+++ crates/fr-command/src/lib.rs
@@ -31381,7 +31381,9 @@ mod tests {
             0,
         )
         .expect("eval should succeed");
-        assert_eq!(reply, RespFrame::Integer(2));
+        // Upstream PROPAGATE_REPL = 2, so REPL_REPLICA + REPL_SLAVE = 4.
+        // (br-frankenredis-replconst)
+        assert_eq!(reply, RespFrame::Integer(4));
 
         let actual: Vec<(Vec<Vec<u8>>, u8)> = store
             .script_propagation_records
